backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.32
sqlmodel==0.0.21
aiosqlite==0.20.0
python-multipart==0.0.9
pdfminer.six==20240706
python-docx==1.1.2
prometheus-client==0.20.0
structlog==24.2.0
orjson==3.10.7
httpx==0.27.0
pyjwt==2.9.0
pytest==8.3.2
requests==2.32.3
greenlet>=3.0,<4.0
